Simplify Poisson Binomial distribution
In Ex Nihilo, rewards from hammering are expressed as floating point
probabilities, while rewards from sieving are expressed as integer
rarities (essentially odds). I tried to unify these two in order to
simplify, however, this complicated the distribution (for calculating
probabilities).

There is a very simple relation between the probabilities and the odds,
that is however susceptible to a floating point error. In my
investigations, converting an int to a float, and back to an int, is
safe for all non-zero integers greater than -5939742 and less than
5933443. So for all practical purposes it won't be a problem.