Make the PMF a list of pairs instead of map
When making use of the Poisson Binomial distribution PMF, I need access
to both the integer input and it's corresponding probability; an
excellent use case for a map. However, I also found that I need to be
able to access elements by an index; an excellent use case for a list.
So I compromised and went with a list of pairs (or 2-tuples).

I'm not necessarily convinced that a pair type is really ideal, but it's
quick and easy and I already had a generic implementation. I really
should probably be using specific types, but naming things is one of the
two hard things in computer science (the others being cache
invalidation, and off by one errors).