Delay building distribution; fixes #2
If we wait to build the probability distribution, then we can modify the
probabilities before we calculate the PMF. In particular, this allows us
to display the correct outcomes for a fortune enchanted hammer.
Factoring in the luck multiplier, exposed a minor bug, namely that
probabilities should never be greater than 1, so we take that into
account when calculating the PMF.

We also switch to displaying odds against, instead of odds for in order
to make it more obvious that the numbers are odds and not probabilities.