Setup libs and build.properties
The Chicken Bones libraries can be gotten automatically from their ivy
repository; we now make use of that. We also emulate maven's provided
scope for dependencies, for those which should be in the compile time
classpath, but not the runtime classpath.

We also add a build.properties configuration for specifying some version
numbers.